+++ src/components/VideoPopup.jsx
@@ -1,241 +1,151 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+/* eslint-disable react/prop-types */
+import React, { useEffect, useRef, useState } from 'react';
+import { Radio, Input, Modal, Button } from 'antd';
 import IPFSService from '../services/IPFSService';
 import config from '../configuration';
-import ErrorPopup from './ErrorPopup';
 
-const Label = ({ children }) => (
-  <span className="label" style={{ marginRight: '0.5rem' }}>
-    {children}
-  </span>
-);
+const VideoPopup = ({ visible, handleClose, reactQuillRef }) => {
+  const [type, setType] = useState(1);
+  const [url, setURL] = useState('');
+  const [currentState, setCurrentState] = useState('');
+  const [loading, setLoading] = useState(false);
+  const stream = useRef(null);
+  const file = useRef(null);
 
-Label.propTypes = {
-  children: PropTypes.node.isRequired,
-};
+  const onChange = e => {
+    setType(e.target.value);
+  };
 
-function convertVideoUrlToEmbeded(url) {
-  let match =
-    url.match(/^(https?):\/\/(?:(?:www|m)\.)?youtube\.com\/watch.*v=([a-zA-Z0-9_-]+)/) ||
-    url.match(/^(https?):\/\/(?:(?:www|m)\.)?youtu\.be\/([a-zA-Z0-9_-]+)/);
-
-  if (match) {
-    return `${match[1]}://www.youtube.com/embed/${match[2]}?showinfo=0`;
-  }
-
-  // eslint-disable-next-line no-cond-assign
-  if ((match = url.match(/^(https?):\/\/(?:www\.)?vimeo\.com\/(\d+)/))) {
-    return `${match[1]}://player.vimeo.com/video/${match[2]}/`;
-  }
-  return url;
-}
-
-class Content extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      type: null,
-      url: null,
-      file: null,
-      blob: null,
-      currentState: '',
-      stream: null,
-      recordVideo: null,
-    };
-
-    this.handleFile = this.handleFile.bind(this);
-    this.handleCamera = this.handleCamera.bind(this);
-    this.handleUpload = this.handleUpload.bind(this);
-    this.startRecording = this.startRecording.bind(this);
-    this.stopRecording = this.stopRecording.bind(this);
-    this.detectExtension = this.detectExtension.bind(this);
-    this.handleScreenSharing = this.handleScreenSharing.bind(this);
-  }
-
-  handleFile(e) {
-    this.setState({
-      file: URL.createObjectURL(e.target.files[0]),
-      blob: e.target.files[0],
-    });
-  }
-
-  handleCamera() {
-    this.setState({ type: 'camera', currentState: '' }, () => {
-      navigator.getUserMedia(
-        { audio: true, video: true },
-        cameraStream => {
-          this.setState(
-            {
-              stream: new window.MultiStreamsMixer([cameraStream]),
-              cameraStream,
-            },
-            () => {
-              this.state.stream.frameInterval = 1; // eslint-disable-line react/no-direct-mutation-state
-              this.state.stream.startDrawingFrames();
-              window.setSrcObject(
-                this.state.stream.getMixedStream(),
-                document.getElementById('video'),
-              );
-            },
-          );
-        },
-        () => {
-          alert('No camera devices found');
-        },
-      );
-    });
-  }
-
-  handleExit() {
-    const { cameraStream, audioStream, screenStream } = this.state;
-    if (cameraStream) {
-      cameraStream.stop();
-      cameraStream.getTracks()[0].stop();
-      cameraStream.getTracks()[1].stop();
-    }
-    if (audioStream) {
-      audioStream.stop();
-      audioStream.getTracks()[0].stop();
-    }
-    if (screenStream) {
-      screenStream.stop();
-      screenStream.getTracks()[0].stop();
-    }
-  }
-
-  handleUpload() {
-    this.setState(
-      {
-        currentState: 'uploading',
+  const handleCamera = () => {
+    navigator.getUserMedia(
+      { audio: true, video: true },
+      cameraStream => {
+        stream.current = new window.MultiStreamsMixer([cameraStream]);
+        stream.current.frameInterval = 1;
+        stream.current.startDrawingFrames();
+        window.setSrcObject(stream.current.getMixedStream(), document.getElementById('video'));
+        // this.setState(
+        //   {
+        //     stream: new window.MultiStreamsMixer([cameraStream]),
+        //     cameraStream,
+        //   },
+        //   () => {
+        //     this.state.stream.frameInterval = 1; // eslint-disable-line react/no-direct-mutation-state
+        //     this.state.stream.startDrawingFrames();
+        //     window.setSrcObject(
+        //       this.state.stream.getMixedStream(),
+        //       document.getElementById('video'),
+        //     );
+        //   },
+        // );
       },
       () => {
-        IPFSService.upload(this.state.blob)
-          .then(hash => {
-            this.handleExit();
-
-            this.props.handleQuillInsert(config.ipfsGateway + hash.slice(6));
-          })
-          .catch(err => {
-            ErrorPopup('Something went wrong with the upload.', err);
-            console.log(err);
-          });
+        alert('No camera devices found');
       },
     );
-  }
+  };
 
-  startRecording() {
-    const { stream } = this.state;
-    this.setState({ currentState: 'recording' });
-    window.setSrcObject(stream.getMixedStream(), document.getElementById('video'));
-    this.setState(
-      {
-        recordVideo: window.RecordRTC(stream.getMixedStream(), {
-          type: 'video',
-        }),
-      },
-      () => this.state.recordVideo.startRecording(),
-    );
-  }
-
-  stopRecording() {
-    const { recordVideo } = this.state;
-    recordVideo.stopRecording(() => {
-      this.setState({
-        file: window.URL.createObjectURL(recordVideo.getBlob()),
-        blob: recordVideo.getBlob(),
-        currentState: 'recording-stopped',
-      });
-    });
-  }
-
-  detectExtension() {
+  const detectExtension = () => {
     const extensionid = 'ajhifddimkapgcifgcodmmfdlknahffk';
     const image = document.createElement('img');
     image.src = `chrome-extension://${extensionid}/icon.png`;
 
     image.onload = () => {
-      this.handleScreenSharing();
+      // handleScreenSharing();
     };
     image.onerror = () => {
-      this.setState({ currentState: 'missing extension' });
+      setCurrentState('missing extension');
     };
+  };
+
+  useEffect(() => {
+    if (type === 3) {
+      handleCamera();
+    } else if (type === 4) {
+      detectExtension();
+    }
+  }, [type]);
+  function clearStates() {
+    if (file.current) {
+      file.current.value = null;
+    }
+    setLoading(false);
+    setURL(false);
   }
 
-  handleScreenSharing() {
-    this.setState({ type: 'screen', currentState: '' });
-    window.getScreenId((err, sourceId, screenContraints) => {
-      if (err) {
-        console.log(err);
-      }
-      navigator.mediaDevices.getUserMedia({ video: false, audio: true }).then(audioStream => {
-        navigator.getUserMedia = navigator.mozGetUserMedia || navigator.webkitGetUserMedia;
-        navigator.mediaDevices.getUserMedia(screenContraints).then(screenStream => {
-          screenStream.fullcanvas = false;
-          screenStream.width = window.screen.width; // or 3840
-          screenStream.height = window.screen.height; // or 2160
-          this.setState(
-            {
-              stream: new window.MultiStreamsMixer([screenStream, audioStream]),
-              audioStream,
-              screenStream,
-            },
-            () => {
-              this.state.stream.frameInterval = 1; // eslint-disable-line react/no-direct-mutation-state
-              this.state.stream.startDrawingFrames();
-              window.setSrcObject(
-                this.state.stream.getMixedStream(),
-                document.getElementById('video'),
-              );
-            },
-          );
+  function closeModal() {
+    clearStates();
+    handleClose();
+  }
+
+  function insertToEditor(videURL) {
+    const quill = reactQuillRef.current.getEditor();
+    const index = quill.getLength() - 1;
+    quill.insertEmbed(index, 'video', videURL);
+  }
+
+  function uploadVideoToIPFS() {
+    const reader = new FileReader();
+    reader.onload = _ => {
+      IPFSService.upload(file.current.files[0])
+        .then(hash => {
+          insertToEditor(config.ipfsGateway + hash.slice(6));
+          closeModal();
+        })
+        .catch(err => {
+          // ErrorPopup('Something went wrong with the upload.', err);
+          console.log(err);
+        })
+        .finally(() => {
+          setLoading(false);
         });
-      });
-    });
+    };
+    setLoading(true);
+    reader.readAsDataURL(file.current.files[0]);
   }
 
-  render() {
-    const { handleQuillInsert } = this.props;
-    const { type, url, file, currentState } = this.state;
-    return (
-      <div>
-        <span className="label" style={{ display: 'block', marginBottom: '1rem' }}>
-          Choose type of video
-        </span>
-        <Label>
-          <input name="type" type="radio" onClick={() => this.setState({ type: 'link' })} /> Link
-        </Label>
-        <Label>
-          <input name="type" type="radio" onClick={() => this.setState({ type: 'file' })} /> File
-        </Label>
-        <Label>
-          <input name="type" type="radio" onClick={() => this.handleCamera()} /> Camera
-        </Label>
-        <Label>
-          <input name="type" type="radio" onClick={() => this.detectExtension()} /> Screen sharing
-        </Label>
-        <div style={{ marginTop: '1rem' }}>
-          {type === 'link' && (
-            <div>
-              <input
-                style={{ width: '100%', marginBottom: '1rem' }}
-                type="text"
-                placeholder="Insert video URL"
-                onChange={e => this.setState({ url: e.target.value })}
-              />
-              <button
-                type="button"
-                onClick={() => handleQuillInsert(convertVideoUrlToEmbeded(url))}
-              >
-                Submit
-              </button>
-            </div>
-          )}
-          {type === 'file' && (
-            <div style={{ display: 'flex', justifyContent: 'center' }}>
-              <input type="file" accept="video/*; image/*" onChange={this.handleFile} />
-            </div>
-          )}
-        </div>
+  function onOk() {
+    switch (type) {
+      case 1:
+        insertToEditor(url);
+        closeModal();
+        break;
+      case 2:
+        uploadVideoToIPFS();
+        break;
+      default:
+        break;
+    }
+  }
+
+  return (
+    <Modal
+      title="Attach a video to description"
+      visible={visible}
+      onOk={onOk}
+      onCancel={closeModal}
+      footer={[
+        <Button key="back" onClick={closeModal}>
+          cancel
+        </Button>,
+        <Button key="submit" type="primary" loading={loading} onClick={onOk}>
+          Submit
+        </Button>,
+      ]}
+    >
+      <Radio.Group onChange={onChange} value={type}>
+        <Radio value={1}>Link</Radio>
+        <Radio value={2}>File</Radio>
+        {/* <Radio value={3}>Camera</Radio> */}
+        {/* <Radio value={4}>Screen sharing</Radio> */}
+      </Radio.Group>
+      <div style={{ paddingTop: '40px' }}>
+        {type === 1 && <Input placeholder="Video URL" onChange={e => setURL(e.target.value)} />}
+        {type === 2 && (
+          <div style={{ display: 'flex', justifyContent: 'center' }}>
+            <input type="file" accept="video/*" ref={file} />
+          </div>
+        )}
         {currentState === 'missing extension' && (
           <div role="alert">
             <strong>You need to install this </strong>
@@ -249,67 +159,9 @@ class Content extends Component {
             <strong>and reload</strong>
           </div>
         )}
-        {(file || type === 'camera' || type === 'screen') && (
-          <div>
-            <video
-              style={{
-                marginTop: '1rem',
-                width: '100%',
-                display: currentState !== 'recording-stopped' ? '' : 'none',
-              }}
-              controls
-              muted
-              autoPlay
-              id="video"
-            />
-            {currentState === 'recording-stopped' && (
-              <video style={{ marginTop: '1rem', width: '100%' }} controls autoPlay src={file} />
-            )}
-            {(type === 'camera' || type === 'screen') && (
-              <div style={{ marginTop: '1rem', marginBottom: '1rem' }}>
-                <button
-                  type="button"
-                  className={currentState === 'recording' ? 'btn btn-danger' : 'btn btn-success'}
-                  onClick={() =>
-                    currentState === 'recording' ? this.stopRecording() : this.startRecording()
-                  }
-                >
-                  {currentState === 'recording' ? 'Stop recording' : 'Start recording'}
-                </button>
-              </div>
-            )}
-            <button type="button" className="btn btn-info" onClick={this.handleUpload}>
-              {currentState === 'uploading' ? 'Uploading...' : 'Upload'}
-            </button>
-          </div>
-        )}
-        <button
-          type="button"
-          style={{ marginTop: '1rem' }}
-          className="btn btn-danger"
-          onClick={() => {
-            this.handleExit();
-            React.swal.close();
-          }}
-        >
-          Close
-        </button>
       </div>
-    );
-  }
-}
-
-Content.propTypes = {
-  handleQuillInsert: PropTypes.func.isRequired,
+    </Modal>
+  );
 };
 
-export default handleQuillInsert => {
-  React.swal({
-    title: 'Attach a video to Milestone',
-    content: React.swal.msg(<Content handleQuillInsert={handleQuillInsert} />),
-    button: {
-      visible: false,
-    },
-    closeOnClickOutside: false,
-  });
-};
+export default VideoPopup;
